poulateFromCSV method moved to the class CSVDataReader
